'com.addthis.metrics' -> 'com.addthis.metrics3'
In the reporter-config3 module rename the package so that it doesn't
collide with the package name of the reporter-config2 module.
