+++ reporter-config-base/src/main/java/com/addthis/metrics/reporter/config/AbstractConsoleReporterConfig.java
@@ -21,7 +21,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 
-class AbstractConsoleReporterConfig extends AbstractMetricReporterConfig
+public class AbstractConsoleReporterConfig extends AbstractMetricReporterConfig
 {
     private static final Logger log = LoggerFactory.getLogger(AbstractConsoleReporterConfig.class);
