+++ reporter-config3/src/test/java/com/addthis/metrics3/reporter/config/sample/SampleTest.java
@@ -12,9 +12,9 @@
  * limitations under the License.
  */
 
-package com.addthis.metrics.reporter.config.sample;
+package com.addthis.metrics3.reporter.config.sample;
 
-import com.addthis.metrics.reporter.config.ReporterConfig;
+import com.addthis.metrics3.reporter.config.ReporterConfig;
 
 import com.codahale.metrics.Counter;
 import com.codahale.metrics.Meter;
